app.py
import streamlit as st
import speech_recognition as sr
import pandas as pd
import matplotlib.pyplot as plt
from google.cloud import speech
from google.oauth2 import service_account
import io
import threading
import time
import random  # For simulating heart rate data

# Initialize speech recognition
recognizer = sr.Recognizer()

# Simulate Bluetooth connection and heart rate monitoring
def simulate_heart_rate():
    while True:
        yield random.randint(60, 100)
        time.sleep(1)

# Function to check if heart rate is elevated
def is_heart_rate_elevated(rate):
    return rate > 90

# Function for real-time transcription
def transcribe_audio(audio_file):
    # In a real implementation, you would use Google Cloud Speech-to-Text API here
    # For this example, we'll use a placeholder transcription
    return "This is a placeholder transcription of the conversation."

# Main app
def main():
    st.title("Couple's Conversation Navigator")

    # Sidebar for user information and settings
    st.sidebar.header("User Settings")
    user1_name = st.sidebar.text_input("User 1 Name")
    user2_name = st.sidebar.text_input("User 2 Name")

    # Main conversation area
    st.header("Start Your Conversation")
    st.write("Use this space to have your conversation. Remember to be respectful and listen to each other.")

    # Simulated heart rate monitoring
    if st.button("Start Heart Rate Monitoring"):
        heart_rate_data = {"User 1": [], "User 2": []}
        
        for i in range(10):  # Simulate 10 seconds of heart rate data
            user1_rate = next(simulate_heart_rate())
            user2_rate = next(simulate_heart_rate())
            
            heart_rate_data["User 1"].append(user1_rate)
            heart_rate_data["User 2"].append(user2_rate)
            
            if is_heart_rate_elevated(user1_rate):
                st.warning(f"{user1_name}'s heart rate is elevated. Take a deep breath and try to calm down.")
            if is_heart_rate_elevated(user2_rate):
                st.warning(f"{user2_name}'s heart rate is elevated. Take a deep breath and try to calm down.")
            
            time.sleep(1)
        
        # Display heart rate chart
        df = pd.DataFrame(heart_rate_data)
        st.line_chart(df)

    # Voice recording
    st.header("Record Your Conversation")
    if st.button("Start Recording"):
        with sr.Microphone() as source:
            st.write("Recording... Speak now.")
            audio = recognizer.listen(source, timeout=10)
            st.write("Recording complete.")
        
        # Transcribe audio
        transcription = transcribe_audio(audio)
        st.subheader("Transcription")
        st.write(transcription)

    # Conversation guidance
    st.header("Conversation Guidance")
    guidance = [
        "Use 'I' statements to express your feelings.",
        "Listen actively without interrupting.",
        "Acknowledge your partner's perspective.",
        "Focus on the issue at hand, not past grievances.",
        "Take breaks if the conversation becomes too heated.",
    ]
    for tip in guidance:
        st.markdown(f"- {tip}")

    # Summary feature
    st.header("Conversation Summary")
    summary = st.text_area("Enter key points from your conversation:")
    if st.button("Generate Summary"):
        st.subheader("Areas of Agreement")
        st.write("Placeholder for areas of agreement")
        st.subheader("Areas of Disagreement")
        st.write("Placeholder for areas of disagreement")
        st.subheader("Action Items")
        st.write("Placeholder for action items")

if __name__ == "__main__":
    main()