requirements.txt
streamlit==1.12.0
SpeechRecognition==3.8.1
pandas==1.3.5
matplotlib==3.5.2
google-cloud-speech==2.13.1